refactor: Switch away from for-loops
This was inspired by pushgen, see
https://github.com/joaquintides/transrangers/blob/master/rust.md

As expected, this wasn't actually a performance win.  Its believed to be
that way because we iterators aren't specializing their `try_for_each`
impl because `Try` isn't stablized to do so.

When running `cargo bench --bench checks`, performance was hit or miss.
For some data sets, performance varied from no change, to +-4%,to +-8%.
There wasn't anything to predice it, so unsure if that is just natural
variance between benchmark runs.